Add very barebones error and index pages
Both pages were built as streamlit scripts for the prototype as doing
things this way allowed us to dodge the need to properly build a UI
while still getting something functional. In the hardened version of
this feature, we'll probably want to build special pages for both of
these in the frontend.

The index page currently doesn't actually provide an index to the user
-- we'll be implementing this in the next commit.